feat: share validation utils and add debug options
