+++ scripts/check_accuracy.py
@@ -18,6 +18,7 @@ if not HAS_PDFPLUMBER:
     print("pdfplumber not installed; using text fallback")
 
 from statement_refinery import pdf_to_csv  # noqa: E402
+from statement_refinery.validation import extract_total_from_pdf
 
 # Golden CSVs checked into ``tests/data`` are generated by running
 # ``pdf_to_csv`` on each statement PDF with ``--out``.  When a new PDF is added
@@ -26,36 +27,6 @@ from statement_refinery import pdf_to_csv  # noqa: E402
 DATA_DIR = ROOT / "tests" / "data"
 
 
-def extract_total_from_pdf(pdf_path: Path) -> Decimal:
-    """Return the total amount printed in the PDF or snapshot text."""
-    txt_path = pdf_path.with_suffix(".txt")
-    if txt_path.exists():
-        text = txt_path.read_text()
-    elif HAS_PDFPLUMBER:
-        import pdfplumber
-
-        with pdfplumber.open(str(pdf_path)) as pdf:
-            text = "\n".join(page.extract_text() or "" for page in pdf.pages)
-        txt_path.write_text(text)
-    else:
-        raise FileNotFoundError(f"No text fallback for {pdf_path.name}")
-
-    patterns = [
-        r"Total desta fatura\s*[=R\$\s]*([\d\.]+,\d{2})",
-        r"Total da fatura\s*[=R\$\s]*([\d\.]+,\d{2})",
-        r"Total\s*[=R\$\s]*([\d\.]+,\d{2})",
-        r"= Total desta fatura\s*[=R\$\s]*([\d\.]+,\d{2})",
-        r"TOTAL\s*[=R\$\s]*([\d\.]+,\d{2})",
-        r"Valor Total\s*[=R\$\s]*([\d\.]+,\d{2})",
-        r"Saldo Total\s*[=R\$\s]*([\d\.]+,\d{2})",
-    ]
-    for pattern in patterns:
-        match = re.search(pattern, text)
-
-        if match:
-            val = match.group(1).replace(".", "").replace(",", ".")
-            return Decimal(val)
-    raise ValueError(f"Could not find total in {pdf_path.name}")
 
 
 def compare(
